fix compilation error caused by int overflow on some platform (#12501)

+++ pkg/security/secl/model/consts_linux_arm64.go
@@ -22,5 +22,5 @@ var (
 		"PTRACE_SYSEMU_SINGLESTEP": unix.PTRACE_SYSEMU_SINGLESTEP,
 	}
 
-	mmapFlagArchConstants = map[string]int{}
+	mmapFlagArchConstants = map[string]uint64{}
 )
